+++ R/data.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #'
-#' This data is taken from \pkg{RecordLinkage} R package developed by Murat Sariyar and Andreas Borg.
+#' This data is taken from the RecordLinkage R package developed by Murat Sariyar and Andreas Borg.
 #' The package is licensed under GPL-3 license.
 #'
 #' The `RLdata500` table contains artificial personal data.