+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # version 1.0.3
 
++ Added support for the `on` and `on_blocking` arguments
+in the `blocking()` and `pair_ann()` functions.
 + Fixed errors in the evaluation functions.
 
 # version 1.0.2